refactor: 프로젝트명, 디렉토리명 변경에 따른 내부 코드 수정
