+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'oraclejpa'
+rootProject.name = 'boardspace'
