+++ src/main/java/com/boardspace/repository/UserRepository.java
@@ -1,15 +1,10 @@
-package com.oraclejpa.repository;
+package com.boardspace.repository;
 
-import com.oraclejpa.model.Post;
-import com.oraclejpa.model.User;
+import com.boardspace.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
