+++ src/main/java/com/boardspace/service/PostService.java
@@ -1,7 +1,7 @@
-package com.oraclejpa.service;
+package com.boardspace.service;
 
-import com.oraclejpa.model.Post;
-import com.oraclejpa.repository.PostRepository;
+import com.boardspace.model.Post;
+import com.boardspace.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
