+++ src/test/java/com/boardspace/BoardspaceApplicationTests.java
@@ -1,10 +1,10 @@
-package com.oraclejpa;
+package com.boardspace;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class OraclejpaApplicationTests {
+class BoardspaceApplicationTests {
 
 	@Test
 	void contextLoads() {
